Allow both member and root for `cpuset.cpus.partition`.
Previously we used `root` for `cpuset.cpus.partition` of slices, to
exclude its cpus from the parent partition.

To workaround a kernel bug we will try using `member` to not exclude
cpus from the parent.

See the next commit for details.

Since we will have a mix of slices with `root` (from before) and
`member` (new ones) values, this commit changes the availability check
to take that into account.