Monitor: only increment checkup when not already set
Failed monitoring jobs incremented the checkup semaphore whenever the
resource responded to `incr_checkup`. In production we had a case where
a resource accumulated almost 150k checkup semaphores. In order to avoid
runaway accumulation of these semaphores, we now only increment the
checkup semaphore when the resource supports it and has not already set
a checkup .