Do not create firewalls for subnets in general
The default subnet included with a project defaults to all-open to
ease the learning curve.

However, all subnets in general follow the same algorithm: if a
firewall is not passed, it gins up a new firewall record with all-open
rules.  This is dangerous.  As-is the only way to suppress this
implied firewall creation is to get the create request on private
subnet exactly right, and pass a firewall object from the get-go.
Failing (or not knowing) this, you have to read back the subnet to go
and delete the all-open firewall rule.  I, personally, totally forgot
about how to suppress the default rules until reminded.