Fix synchronization issue for connected subnets and nic adding operation
This commit fixes a bug where we hit when we create multiple subnets
with a single VM each and connecting them. Before this commit, since we
depended on "rekey_payload" field to be nil or not while filtering on
rekeying nics, we made the  mistake of working on nics that are not yet
fully added to our sister subnets. They would already be in the process
of just being added to their own subnet and we would falsely assume that
they are part of the refresh_keys workstream of another subnet. With
this commit, we store the nics that we work on on our strand stack and
in later states, we resolve to only those.