Start using the newly added state and update it in NicNexus.rb
SubnetNexus/NicNexus relationship has been quite fragile. The dependency
on the semaphores + strand label proved to be quite useless. Especially
considering the label names may change, or new labels may be introduced.
Recently we had an issue of IpsecTunnels not being created for a VM that
was added to the subnet. When we add the new nic, or rekeying from
multiple different "connected" subnets, and no control over who does
what/when, this caused an issue. One of the nics were in the process of
getting added while the others were getting rekeyed. This caused the
label based filtering to completely fail. Instead, with this commit, we
are introducing the entity level state where we control every step and
filter using that state in a later commit.