Fix Subnet AZ pickup and controlplane records mixup
We used to call az_to_provision_subnet which picks a subnet az among 3
available "a", "b" and "c" if not a specific one is specified. Once we
use that picked az to provision the subnet, while making the update to
the nic_aws_resource entity, we were recalling the function. It results
in a mismatch.