Use random file names for write-read healthchecks of VmHost
There were times where we would run VmHost healthchecks in parallel,
and the code prior to this commit had a race condition for reading or
deleting the file, causing the healthcheck to randomly fail.

With this commit, every time we run the healthcheck, we are ensuring a
unique file name to avoid the race condition.