Prevent deletion of GitHub installations for suspended accounts
When we suspend an account, the user can’t log in to the Ubicloud
console, which prevents them from removing the GitHub installation. This
helps block fraudulent users from repeatedly reusing the same GitHub
organization.

However, if they remove the installation from the GitHub UI, we receive
a webhook event and currently delete the installation on our side. This
allows them to reuse the same GitHub organization with a new account.

To prevent this, if the installation’s project is inactive, we no longer
delete the installation when handling the webhook event.