Fix lease violations in LoadBalancerNexus#destroy
`strand.children.each(&:destroy)` and variants of it are not safe and can
cause lease violations. Outside of test code, there are only two places that a
strand can be safely destroy:

* `Prog::Base#reap` (parents reaping exited child strands)
* `Strand#unsynchronized_run` (self-reaping exited strands without parents)

Any other case where strands are being destroyed is probably a bug that could
cause a lease violation, since it is possible the strand is leased at the time
it is destroyed.

Fix this case by supporting the destroy semaphore in both child strands (which
are either CertServer or UpdateLoadBalancerNode). If the destroy semaphore is
set, have the strands pop. Add LoadBalancerNexus#wait_children_destroyed and
have it reap the child strands before adding strands to remove the VMs from the
load balancer and hop to wait_all_vms_removed.

While here, in before_run, replace wait_destroy with wait_all_vms_removed. This
has been broken since wait_destroy was renamed to wait_all_vms_removed in
fa350fa54d2d7abcd49fd2dd5093251406999211. Yet more evidence that the label name
checking approach in before_run is too brittle and should be changed.