Fix lease violations in VictoriaMetricsServerNexus#destroy
Use a similar approach as for LoadBalancerNexus#destroy. However,
this is a bit more complicated because VictoriaMetricsServerNexus
can bud an instance of itself for restart, in addition to pushing a
frame for restart. Handle this in before_run by exiting if
in restart state with a parent strand. If in restart state and
there is no parent strand, then hop from restart to destroy.