Fix lease violations in PostgresResourceNexus#destroy
Similar approach as to that used in VictoriaMetricsServerNexus#destroy,
though instead of restart, it is the trigger_pg_current_xact_id_on_parent
label that needs handling in before_run, and it's a bit simpler, as that
is only budded and not pushed or hopped to.

As PostgresResourceNexus can bud ConvergePostgresResource,
ConvergePostgresResource needs to support the destroy semaphore.