Heuristic overcommit for PG instances
This commit adds a heuristic to calculate the overcommit value for PG
for PG instances.

Based on observations from allocation failure related pages, we setup a
y = mx + c type value for overcommit_kbytes, which gives a fixed
overhead of 256Mi (for an extra bump to the smaller SKUs) and a factor
of 75% over the actual available pages that can be charged.