Use PATCH for updating firewall description
This is more consistent with other API endpoints, even though it
currently only allows changing a single field.

This uses r.api_patch_web_post so that the web still uses POST,
allowing for normal Capybara testing and not requiring any spec
changes for the web. This required moving the route above the
r.is call.