Do not use the reserved ips of Vm range for kubernetes nodes
Kubernetes Vms have bigger subnets and each gets a /26 ipv4 range
in order to provision pods using the CNI.

Prior to this commit, we were using the vm_net4.nth(1) for the Vm IP
which is wrong and the first 4 ips should not be used.

First ip of the range is the network address
Second ip should be presumed as gateway
The third and fourth ips are reserved for future proofing.

For the fix, we would call the vm_net4.nth(4) when cidr is not /32.