Fix earliest restore time test
We recently added earliest restore time to the CLI output in tests in
390f2d.

The PR passed CI when merged, but the test started failing after 8 days
because BACKUP_BUCKET_EXPIRATION_DAYS is set to 8.

After that window, the earliest restore time becomes empty, while the
test still expects it to be non-empty.