Add a "limited" project reputation level
Currently, we have two project reputation levels: verified and new. new
is the default and has generous quotas compared to verified.

To reduce fraud, we introduce a new limited level with lower quotas than
new. Projects are assigned limited if their email domain is Gmail. We
can improve this in the future by adding more generic email services to
the list or other heuristics. This helps cover a significant portion of
potential fraud cases.

If the account has any verified projects, the reputation is not
downgraded.