Add NetSsh::Safety utility module
This module is designed to address the potentially problematic usage
of string interpolation to construct SSH command strings. We recently
found a case where an interpolated string was user-modifiable and not
shell escaped. While that is potentially the only case, in general,
the approach is risky.  Instead of an opt-in shell escaping approach,
we should use an opt-out shell escaping approach, in which shell
escaping is always performed unless specifically requested not to.
This is similar to how our HTML escaping works for our templates,
where `<%=` escapes, and you must use `<%==` if you do not want HTML
escaping.

Assuming that argument is derived from user input, instead of
using interpolation:

```ruby
sshable.cmd("command with #{argument}")
```

You use :placeholders in the command string, with values
provided by keyword arguments:

```ruby
sshable.cmd("command with :argument", argument:)
```

All placeholder subsitutions are shell escaped. Note that
substitutions are based on the provided arguments, so:

```ruby
sshable.cmd(":foo :bar", foo: "baz quux")
```

In rare cases, it can be useful to pass an array of values and use
shelljoin, where the arguments of an array are each escaped, but
then joined (with space by default). You can use this by having
your placeholder start with shelljoin_:

```ruby
cmd("command with :shelljoin_arguments", shelljoin_arguments: array)
```