Don't show clover-production in staging pry
This gives a false sense of urgency for staging. We should reserve
this display only for the production environment.  The staging
environment uses RACK_ENV=production to be as close as possible to
the production. Support a STAGING=true environment variable that
the staging environment can also use, and suppress the
clover-production display in pry if set. We could potentially use
Config.staging in additional cases in the future.