Reintroduce code & conflations to model/postgres_timeline.rb
I took this snapshot direct from origin/main, and it made many tests
pass.

The code i peppered with ternary operators for Postgres on AWS
vs. other. That's obviously going to get pretty ugly if we ever add a
third, or fourth, substrate, so probably a good place to look to do
some software engineering in general.