Reintroduce MinIO location and aws? predicate
I recently decided to re-accept code for postgres_timeline, which is
one of the more entangled cases right now, and that means changing
some test fixtures.  Though not marked, the tests are also
significantly conflated.

MinIO needed a Location association, and the Location code/model is
still pretty much all there and probably ought to be for the long
term.