Add stopped label to postgres server
We need to stop a few postgres resources every month due to fraudulent
activity before permanently deleting the data.

Currently, we destroy the resource, since we can restore from backup if
needed. However, it consumes some engineering time.

We can manually stop the resource on the dataplane, which causes an
unavailability page, and users can't see that it's stopped in the UI.

A proper stop may require releasing CPU and memory resources, but the
initial version only stops the PG process without releasing those
resources.

This will help with operational tasks.