Add Custom/SafeShellCommand RuboCop cop
Run the tests with:

    COVERAGE=rubocop rspec -O /dev/null spec/rubocop/cop/custom/safe_shell_command_spec.rb

This cop detects nearly-unsafe shell command patterns in `cmd()` calls
using Shellwords identity checking. It rewrites certain interpolated
strings into a %W[] construction designed to be passed to shelljoin.
It has a focus on heuristic autocorrection, to spare our fingers.

A simple one might be:

    cmd("ls #{foo}")

Where the code is rewritten to:

    cmd(%W[ls #{foo}])

With the idea that the array should be shelljoined, thus, all its argv
are safely escaped.

And slightly more complex case is:

    cmd("ls directory/#{foo}")

In such a case, it's presumed that `directory/#{foo}` are to be
escaped together:

    cmd(%W[ls directory/#{foo}])

An example of a mistaken one is:

   cmd(%W[ls /vm/#{q_vm}])

Where the `q_` prefix means "quoted for shell".  That could be added, or, perhaps with the convenience of this rewrite, simplifying to:

   cmd(%W[ls /vm/#{vm}])

Will become the new way.

Among the more complex (maybe not even worth it) ones were handling
multi-line continuations:

    @nic.vm.vm_host.sshable.cmd("sudo -- xargs -I {} -- ip -n #{@namespace} xfrm state add " \
      "src #{src} dst #{dst} proto esp spi #{spi} reqid #{@reqid} mode tunnel " \
      "aead 'rfc4106(gcm(aes))' {} 128 #{"sel src 0.0.0.0/0 dst 0.0.0.0/0" if is_ipv4}", stdin: key)

Preserving the line breaks and noting the spaces at the end of lines
to test for safety in newline preservation in %W[]:

    @nic.vm.vm_host.sshable.cmd(%W[sudo -- xargs -I {} -- ip -n #{@namespace} xfrm state add
      src #{src} dst #{dst} proto esp spi #{spi} reqid #{@reqid} mode tunnel
      aead 'rfc4106(gcm(aes))' {} 128 #{"sel src 0.0.0.0/0 dst 0.0.0.0/0" if is_ipv4}], stdin: key)

The tests contain many, many edge cases to reach 100% branch coverage,
plus there are some assertions for topologies that appear unreachable
for now.