Add E2E test for no-HA PG on AWS
Since we have an E2E test for runner on AWS, it’s straightforward to add
a test for no-HA PG on AWS as well.

I didn’t split the metal and AWS tests into separate files because, for
both runners and no-HA PG, the only difference is the arguments passed
to nexus assemble. The overall test logic remains the same.