Increase k8s subnet size to /16
Currently we are provisioning subnets with /18 subnets and each vm gets an
18+8 (/26) ipv4 range, meaning each vm has 2^6 ips.

Each pod consumes two ips from the pool, meaning we can provision 31 pods at
most in a node, which is below the kubelet default (110).

By increasing the subnet size to /16, each vm gets 254 ips for pods, allowing
us to provision 127 pods, which is above kubelet defaults.