Avoid interpolated command string in MetricsTargetMethods
Have the exec! method handle all escaping. Do not require shellwords
to use any of its methods directly.

Change metrics_dir to return an unesacped version of the directory
instead of the escaped version, as the exec! method now handles
escaping. Make the method private to ensure nothing else is calling it.