Avoid interpolated command string in Prog::DnsZone::SetupDnsServerVm
This changes the inhost_name usage to be outside the single quotes,
because shellescaped output should be outside quotes. However, due
to shell parsing, it will be considered part of the same sed
argument. Note that if the inhost_name had any regexp metacharacters,
those wouldn't be escaped. However, since the inhost_name is only
the first 8 characters of the ubid, it shouldn't have either
shell metacharacters or regexp metacharacters.