Avoid interpolated command string in Prog::Github::GithubRunnerNexus
Because of the removal of comments and empty lines, the string is
no longer frozen. Pass it as a stdin script to bash to work around
that. To ensure safety, use NetSsh.command and NetSsh.combine to
build the script passed to stdin.

This adds an extra backlash to the serial log name. This will be
removed by the shell's parsing, but is necessary so the substitution
works.