Backfill AZs field if it's not already set
AWS doesn't have the same number of AZs in every region. Knowing they
have at least 3, we were making use of that fact and not using the whole
region sometimes. This information can be fetched from the AWS. In this
commit, we are adding the method to basically return the database stored
AZs information if it is filled, if not, we are fetching it from the
AWS.