Tighten up NetSsh now that spec issues have been fixed
Add NetSsh::MissingMock exception class, and raise it if the
_cmd/_exec!  methods are not mocked. This is to catch places
where we aren't testing what commands are executed in the specs.
I only found one example where we weren't, which was fixed in an
earlier commit.

Add NetSsh::PotentialInsecurity for cases where interpolated
strings are passed to cmd/exec!. Raise this exception instead of
warning, so tests cannot pass with potentially insecure commands.

Raise TypeError instead of RuntimeError if a non-string is provided.

Add covering specs, and only use nocov around the non-test
environment parts.