Make API endpoint descriptions more concise
I was reviewing our API docs to use it and noticed that many endpoint
titles are overly wordy, which makes it harder to scan the list quickly.

For example, since tokens are already project-specific, we can drop
suffixes like "in the project". Similarly, because delete/show endpoints
already require the location parameter, we can remove "in location" from
those titles.