Add semaphore use_old_walg_command
This way, we will simply increment it for the existing postgres
resources and keep them in the old archival method. We will give it a
month or so and try with the new method in production for the new
resources. Later, we'll migrate them to the new image as well.