Observe archival backlog and page if it exceeds threshold (try 2)
Add an SSH call to count the number of WAL files ready to be archived
and raise a Page if it exceeds a threshold. In the previous attempt,
this was added to pulse_check, which caused a problem. Since postgres
server pulse checks require an event loop, running exec! on the same
session is not safe, and can cause state corruption which lead to stuck
SSH sessions. In this attempt, we move the backlog check to the metrics
export path.