Enable backtrace for the vhost-backend systemd units.
This will be very useful when debugging a panic in prod.

With this change, we expect something like the following if any thread
panics.

```
thread 'vring_worker' panicked at
src/block_device/bdev_lazy/metadata/shared_state.rs:56:41:
index out of bounds: the len is 0 but the index is 0
stack backtrace:
   0: __rustc::rust_begin_unwind
             at ./rustc/.../panicking.rs:697:5
   1: core::panicking::panic_fmt
             at ./rustc/.../panicking.rs:75:14
   2: core::panicking::panic_bounds_check
   ...
   7: ubiblk::...::LazyIoChannel::request_stripes_fetched
             at ubiblk/src/block_device/bdev_lazy/device.rs:63:37
   8: <ubiblk::...::LazyIoChannel as ubiblk::...::add_read
             at ubiblk/src/block_device/bdev_lazy/device.rs:182:28
   ...
```