Prepare routes for using POST instead of DELETE for web routes
Similar to using POST instead of PATCH, we want to use POST
instead of DELETE so that we can avoid the need for javascript
and use standard Capybara testing patterns.

We have significantly more r.delete routes than r.patch routes,
so this is a bigger lift for r.delete than for r.patch. However,
the same principles apply.

For r.delete calls outside of r.is blocks, no changes are needed.
For r.delete calls inside r.is blocks, remove the r.is call or
change it to r.on if it takes a matcher, and add true matchers as
appropriate to the routing methods inside.

The token route has an r.is block that requires a more significant
change, since it has code that was inside the r.is block, after the
r.get call, and before the r.post and r.delete calls. Add a
check_tag_modification! to handle the shared logic instead of having
it inline, and have r.post and r.delete call that method.