Add unused_associations_check rake task
This uses Sequel's unused_associations plugin to detect associations
that are not used at all, or that define methods that are not used.
Such methods consume unnecessary memory, resulting in longer full
garbage collections.

Consider the strand and semaphore associations as always used.
Even if the code doesn't use them, these are likely to be used in
pry.