Add ProviderDispatcher plugin
This plugin allows for splitting methods into aws and metal specific
implementations in their own files. The aws implementation goes into
an aws subdirectory with the same filename, and the metal implementation
goes into a metal subdirectory with the same filename.

The aws implementation uses a model::Aws module, and the metal
implementation uses a model::Metal module. For methods that differ
in their aws and metal implementation, the model::Aws module has
a private method with an aws_ prefix, and the model::Metal module
has a private method with a metal_ prefix. ProviderDispatcher will
then create a prefix less method that delegates to one of those
two providers. This approach should be easy to scale to more than
two providers.

Currently, this calls the aws? method on the model instance to
determine which implementation to use. If the model does not define
their own aws? method, ProviderDispatcher provides a default that
uses location.aws?.

Future commits will use this plugin to separate code in the models.