Use ProviderDispatcher in PostgresServer
This adds add_provider_configs and refresh_walg_blob_storage_credentials
methods so they can have separate aws/metal implementations.

This adds an additional_labels local variable because coverage
testing was invalidly flagging the line as not covered.

This modifies the mocking in the specs so the specs continue to pass.