Fix missed value transformation in form updates
AWS resources has dots in their IDs, which does not generate valid syntax when
used as input[value=...] selectors in jQuery. This change makes sure to replace
dots with dashes when generating such selectors. We already do this in the
redrawChildOptions function, but missed it was missed in another place.