Update PG image to optimize postgres_exporter
In its default configuration, postgres_exporter outputs a number of
timeseries whose cardinality scales with the count of logical databases
on the instance. This produces two bad effects:

1. The metrics scrape size is unbounded, as it scales with however many
   databases the user creates. This leads to a backlog while exporting
   metrics and dropped metrics data on an instance with large (1k+) DBs.
2. More seriously, on low spec instances (standard-2), this can cause
   the exporter process to consume a significant amount of the available
   resources, which can potentially starve other processes disrupting
   other key operations like archiving.

In this change, we disable a lot of the collectors enabled by default.
Additionally, to ensure we still have visibility into some key metrics,
we introduce a set of custom queries that produce instance-aggregated
metrics.