Run GitHub Actions tests on Alpine Linux
We had at least one platform portability issue when using our docker
images, which are built against alpine, and thus busybox and musl
libc.  That was the shebang in
4f6396d82c02f6176639c96f26196910dbc6ecd1. While not caught by the
tests we have right now: it could be, and the first step to automating
catching those regressions is to expand the build matrix.