Add VM host health check to verify clock source is tsc
tsc is the preferred clock source because it is more reliable and
precise. Our hosts use tsc by default, but in some cases they may fall
back to hpet if issues are detected with tsc.

When a host switches to hpet, it can cause instability and runtime
issues for runners. In particular, this leads to errors like "OS
monotonic time stepped backwards!", which we've observed frequently in
Elixir-based workflow jobs.

We should monitor the clock source on VM hosts and alert if any host
switches from tsc to hpet, so we can take corrective action early.