Only consider projects related to the account in github route
This is more hypothetical than actual, since the account must have had
access to the project at one point for the session variable to be set.
However, it's best for consistency to always lookup based on an existing
trusted object.

Reduce mocking in the specs by adding a testing-only
set_github_installation_project_id route.