Implement model hiding in Clover
With model hiding, during coverage testing, access to most models is
proxied, and only specific methods are allowed. This is designed to
prevent future insecure object access vulnerabilities. In general,
object access should be based on associations from existing objects, so
that a logged in account only deals with related projects, and once a
project is selected, the displayed objects (e.g. VMs) must be related to
the project, etc. In general, that's how Clover was already setup, but
with this change, tests fail if an insecure pattern is used.

This whitelists the allowed model calls. Only Account and OidcProvider
are allowed direct (full) access, since those are models used before any
authentication is done.

Because setting this up requires access to all models, it is only done
when autoloads are forced (which coverage testing does). It is not done
in frozen testing mode, or in development or production.