Retry boot image downloads up to 10 times
Currently, if a boot image download fails, we retry indefinitely in
non-production environments and do not retry at all in production.

Additionally, because we raise an exception on failure,
`BootImage.destroy` is never persisted to the database, so the current
behavior is broken.

In production, we usually retry manually after a failure. Instead of
relying on manual retries, we can standardize the behavior across all
environments by retrying the download up to 10 times automatically.