Handle validation failures for PostgreSQL upgrades
Issue found in production Mezmo's missing handle validation failure log,
so this is an actual issue, not just a theoretical problem.

Maybe the error message could be styled better, but if we don't add it,
we only have the flash error, which is going to be cryptic for a user.
FWIW, cryptic flash errors is probably true for most if not all
ValidationFailed exception handling in the web routes.