Improve Vm::Nexus location/nic/subnet validation
This handles the following validation issues that existed:

* You could pick a project-specific location that was tied to another
  project.

* You could pick a subnet that is not in the given location.

* If you know the nic id but no other information, you could gather
  information about it:

  * Whether it has a VM installed
  * Whether it is in the given location
  * Whether it is in the given project

I don't think any of these issues are actual problems, because:

* In routes/project/vm.rb, we call `check_visible_location` to only
  allow locations tied to the project.

* In helpers/vm.rb in `vm_post`, we check the subnet is in the
  given location.

* We don't expose creating a Vm with a specific nic to the user.

So these validations are redundant with previous validations, at least
for web/api requests. However, there may be other internal cases where
this layer of validations could prevent issues. On the flip side, there
may be internal cases where these added validations cause problems
because the code actually wants to create VMs in a project associated
to a location outside the project, or a subnet in a different location.

This allows selecting a project-specific location in a different
project if the VM is allowed to choose a private subnet in a different
location. This is necessary for a couple PostgreSQL specs to pass. I
think it makes sense because the PostgresResource is in a customer
project, in potentially a project-specific location, but it will create
server VMs in the postgres project.

This also changes the `unless nic` into `else`, because if a nic_id is
provided, the code will either skip the `unless nic` branch or raise
an exception.

This adds the Project.nics association to allow only finding nics
related to the project.

This updates the victoria metrics resource prog spec to crate the
resource in the victoria metrics project, instead of a different
project. This is confusing as the same local variable name is used in
the spec to refer to two separate projects.