Add set feature flag action to the admin panel
This new action allows feature flags to be set for projects directly
from the admin panel.

Since feature flags are predefined, a dropdown is used to select the
flag.

The value is optional, if no value is provided, it is treated as nil.

The value is a JSON field, allowing it to be cast to a boolean, integer,
float, array, or hash when possible.