CertNexus.assemble support public+private DNS
This adds an add_private keyword to assemble. When this argument is
used, it will add a "private-" DNS name to the certificate.

This requires rewriting a lot of CertNexus to handle the fact that
there may not be a single DNS challenge, and that you need to add
both challenges to both the public and private names, since I couldn't
find a way to determine which challenge is for which domain name.

This is designed to be used for load balancer SSL certificates.